Added confirmation dialog to delete topic method

+++ src/components/topic.js
@@ -20,10 +20,13 @@ class Topic extends React.Component {
   }
 
   handleDelete = () => {
-    this.props.deleteTopic(this.props.topic.id, this.props.user.token); 
-    this.setState ({
-      complete: true
-    })  
+    const confirmation = window.confirm("Delete this topic?");
+    if (confirmation === true) {
+      this.props.deleteTopic(this.props.topic.id, this.props.user.token); 
+      this.setState ({
+        complete: true
+      })  
+    }
   }
 
   redirectWhenComplete = () => {
